apps/whatsapp-agent-openai/requirements.txt
# WhatsApp Agent Service - Requirements

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2>=3.1.0

# Environment & Config
python-dotenv==1.0.1

# Async & Rate Limiting
aiohttp==3.10.0
aiolimiter==1.1.0

# Google Cloud
google-cloud-firestore==2.18.0
google-cloud-storage>=2.18.0
google-auth>=2.34.0

# AI Provider SDKs
openai>=1.40.0
openai-agents>=0.6.3
anthropic>=0.39.0

# HTTP & Data
requests==2.32.0
httpx>=0.27.0
pydantic>=2.10,<3.0

# Image Processing (optional)
pillow==10.4.0

# Cryptography (WhatsApp Flows encryption)
cryptography>=42.0.0
